+++ src/main/java/nextstep/admin/AdminController.java
@@ -1,34 +1,52 @@
 package nextstep.admin;
 
+import nextstep.common.Authenticated;
+import nextstep.common.LoginMember;
 import nextstep.member.MemberService;
 import nextstep.member.dto.MemberResponse;
+import nextstep.theme.ThemeService;
+import nextstep.theme.dto.ThemeRequest;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
+import java.net.URI;
 import java.util.List;
 
 @RestController
 @RequestMapping("/admin")
 public class AdminController {
     private final MemberService memberService;
+    private final ThemeService themeService;
 
-    public AdminController(MemberService memberService) {
+    public AdminController(MemberService memberService, ThemeService themeService) {
         this.memberService = memberService;
+        this.themeService = themeService;
     }
 
-    @PostMapping("/{id}")
-    public ResponseEntity<Void> upgradeRole(@PathVariable Long id) {
+    @PostMapping
+    public ResponseEntity<Void> registerAdmin(@Authenticated LoginMember loginMember) {
         return null;
     }
 
-    // TODO : 관리자 권한, 경로 같은데 어떻게?
-    @DeleteMapping("/{id}")
-    public ResponseEntity<Void> downgradeRole(@PathVariable Long id) {
+    @DeleteMapping
+    public ResponseEntity<Void> deregisterAdmin(@Authenticated LoginMember loginMember) {
         return null;
     }
 
-    @GetMapping
+    @GetMapping("/members")
     public ResponseEntity<List<MemberResponse>>  getMembers() {
         return null;
+    }
+
+    @PostMapping("/themes")
+    public ResponseEntity<Void> createTheme(@RequestBody ThemeRequest themeRequest) {
+        Long id = themeService.create(themeRequest.toEntity());
+        return ResponseEntity.created(URI.create("/themes/" + id)).build();
+    }
+
+    @DeleteMapping("/themes/{id}")
+    public ResponseEntity<Void> deleteTheme(@PathVariable Long id) {
+        themeService.delete(id);
+        return ResponseEntity.noContent().build();
     }
 }
