+++ src/main/java/nextstep/theme/ThemeController.java
@@ -1,16 +1,10 @@
 package nextstep.theme;
 
-import nextstep.theme.dto.ThemeRequest;
 import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
-import java.net.URI;
 import java.util.List;
 
 @RestController
@@ -22,24 +16,9 @@ public class ThemeController {
         this.themeService = themeService;
     }
 
-    // TODO : admin 권한
-    @PostMapping
-    public ResponseEntity<Void> createTheme(@RequestBody ThemeRequest themeRequest) {
-        Long id = themeService.create(themeRequest.toEntity());
-        return ResponseEntity.created(URI.create("/themes/" + id)).build();
-    }
-
     @GetMapping
     public ResponseEntity<List<Theme>> showThemes() {
         List<Theme> results = themeService.findAll();
         return ResponseEntity.ok().body(results);
-    }
-
-    // TODO : admin 권한
-    @DeleteMapping("/{id}")
-    public ResponseEntity<Void> deleteTheme(@PathVariable Long id) {
-        themeService.delete(id);
-
-        return ResponseEntity.noContent().build();
     }
 }
