+++ src/test/java/nextstep/reservation/ReservationE2ETest.java
@@ -65,7 +65,7 @@ class ReservationE2ETest {
                 .auth().oauth2(accessToken)
                 .contentType(MediaType.APPLICATION_JSON_VALUE)
                 .body(themeRequest)
-                .when().post("/themes")
+                .when().post("/admin/themes")
                 .then().log().all()
                 .statusCode(HttpStatus.CREATED.value())
                 .extract();
